Create WEEK 1
i. Write a program to count number of characters, words, whitespaces and

number of new lines in a given file.
ii. Write a program to count the number of vowels and consonants in a given
string.
iii. Write a program to check whether the entered string is a keyword or not. If
so, then count number of keywords.